requirements.txt
aiofiles==24.1.0
aiohttp==3.12.15
alembic==1.16.5
anthropic==0.34.2
asyncpg==0.30.0
beautifulsoup4==4.14.2
bcrypt==4.3.0
black==25.9.0
celery==5.5.3
cerberus==1.3.7
click==8.2.2
colorlog==6.8.2
configparser==7.1.0
cryptography==46.0.2
diskcache==5.6.3
docker==7.1.0
dynaconf==3.2.11
easyocr==1.7.1
email-validator==2.2.0
fake-useragent==1.5.1
fastapi==0.118.0
flake8==7.3.0
flask==3.1.2
flask-cors==6.0.1
flask-socketio==5.5.1
google-api-python-client==2.147.0
google-auth==2.34.0
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.1
gunicorn==23.0.0
html5lib==1.1
httpx[http2]==0.27.2
imapclient==3.0.1
isort==5.13.2
jinja2==3.1.6
jsonschema==4.25.1
loguru==0.7.3
lxml==5.3.0
marshmallow==4.0.1
mcp==1.16.0
memory-profiler==0.61.0
mypy==1.18.2
nltk==3.9.1
notion-client==2.3.0
numpy==2.2.2
openai==1.102.0
openpyxl==3.1.5
opencv-python-headless==4.12.0.88
pandas==2.3.3
passlib[bcrypt]==1.7.4
pdfplumber==0.11.4
perplexityai==0.13.0
pillow==11.3.0
playwright==1.55.0
prometheus-client==0.23.1
psutil==6.0.0
pyjwt==2.10.1
pylatex==1.4.2
pympler==0.9
pyotp==2.9.0
PyPDF2==3.0.1
pytesseract==0.3.13
pytest==8.4.2
pytest-asyncio==1.2.0
pytest-cov==7.0.0
pytest-mock==3.14.0
pydantic==2.11.0
pydantic-settings==2.11.0
python-dateutil==2.9.0.post0
python-decouple==3.8
python-dotenv==1.1.1
python-jose[cryptography]==3.5.0
python-multipart==0.0.20
pytz==2025.2
pyyaml==6.0.3
redis==5.1.0
reportlab==4.2.5
requests==2.32.3
requests-oauthlib==2.0.0
requests-toolbelt==1.0.0
rich==13.8.1
scrapy==2.13.3
scrapy-splash==0.8.0
scikit-learn==1.7.2
selectolax==0.3.34
selenium==4.25.0
sentence-transformers==3.1.1
spacy==3.8.0
sqlalchemy==2.0.43
starlette==0.48.0
structlog==25.4.0
supervisor==4.2.5
textblob==0.19.0
tiktoken==0.11.0
typer==0.12.5
undetected-chromedriver==3.5.5
uvicorn[standard]==0.31.1
websockets==15.0.1
setuptools>=80.9.0
wheel==0.45.1